Replace all use of ioutil functions in ledger because io/ioutil package is deprecated in Go 1.16.  See
https://go.dev/doc/go1.16#ioutil for more details.

As one example, replace deprecated ioutil.ReadDir() with faster os.ReadDir().

```
name             old time/op    new time/op    delta
IOutilReadDir-4     255µs ± 1%      87µs ± 1%  -65.62%  (p=0.000 n=20+18)

name             old alloc/op   new alloc/op   delta
IOutilReadDir-4    46.2kB ± 0%    20.6kB ± 0%  -55.46%  (p=0.000 n=19+20)

name             old allocs/op  new allocs/op  delta
IOutilReadDir-4       316 ± 0%       216 ± 0%  -31.65%  (p=0.000 n=20+20)
```

Closes #2083